[ICU-3500] Fix some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/15402 Former-commit-id: f0d9c5b6686314a6a76cde1d4a8257af5a1247a8

+++ icu4c/source/test/cintltst/nccbtst.c
@@ -73,8 +73,8 @@ static void setNuConvTestName(const char *codepage, const char *direction)
     sprintf(gNuConvTestName, "[testing %s %s Unicode, InputBufSiz=%d, OutputBufSiz=%d]",
             codepage,
             direction,
-            gInBufferSize,
-            gOutBufferSize);
+            (int)gInBufferSize,
+            (int)gOutBufferSize);
 }
